+++ bindings/python/tests/address_generation_test.py
@@ -6,7 +6,7 @@ import pytest
 import shutil
 
 
-@pytest.mark.skip(reason="temporarily skipped until 2.0 blocks in use")
+@pytest.mark.skip(reason="temporarily skipped")
 def test_address_generation_iota():
     db_path = './test_address_generation_iota'
     shutil.rmtree(db_path, ignore_errors=True)
@@ -28,7 +28,7 @@ def test_address_generation_iota():
     shutil.rmtree(db_path, ignore_errors=True)
 
 
-@pytest.mark.skip(reason="temporarily skipped until 2.0 blocks in use")
+@pytest.mark.skip(reason="temporarily skipped")
 def test_address_generation_shimmer():
     db_path = './test_address_generation_shimmer'
     shutil.rmtree(db_path, ignore_errors=True)
